pyproject.toml
[build-system]
requires = ["pbr>=5.0", "setuptools>=61.0"]
build-backend = "pbr.build"

[project]
name = "cascconf"
dynamic = ["version"]
description = "Cascading Configuration Manager - deep-merge configuration files across multiple directories"
readme = "README.md"
license = {text = "MIT AND GPL-3.0-only"}
requires-python = ">=3.10"
authors = [
    {name = "ConsoleCatzirl"}
]
keywords = [
    "configuration",
    "config",
    "merge",
    "deep-merge",
    "cascading",
    "yaml",
    "json",
    "toml",
    "ini",
    "cli"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
    "Typing :: Typed"
]
dependencies = []

[project.optional-dependencies]
yaml = [
    "pyyaml>=6.0"
]
toml = [
    "tomli-w>=1.0",
    "tomli>=2.0.0; python_version < '3.11'"
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "flake8>=6.0",
    "black>=23.0",
    "isort>=5.12",
    "mypy>=1.0",
    "pre-commit>=3.0",
    "pyyaml>=6.0",
    "types-pyyaml>=6.0",
    "tomli-w>=1.0"
]

[project.scripts]
cascconf = "cascconf.cli:main"

[project.urls]
Homepage = "https://github.com/ConsoleCatzirl/cascconf"
Documentation = "https://github.com/ConsoleCatzirl/cascconf#readme"
Repository = "https://github.com/ConsoleCatzirl/cascconf.git"
Issues = "https://github.com/ConsoleCatzirl/cascconf/issues"
Changelog = "https://github.com/ConsoleCatzirl/cascconf/blob/main/ChangeLog"

[tool.pbr]
warnerrors = true

[tool.black]
line-length = 127
target-version = ["py310", "py311", "py312", "py313", "py314"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 127
known_first_party = ["cascconf"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip_gitignore = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning"
]

[tool.coverage.run]
source = ["cascconf"]
branch = true
omit = [
    "*/tests/*",
    "*/__main__.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:"
]
fail_under = 90
show_missing = true
